pyproject.toml
[tool.poetry]
name = "solomon"
version = "0.0.1"
description = "Project for learning fastapi, services and to build a custom app able to be modified when starting a new project"
license = "MIT"
authors = ["Cristi Ciortea <cristiciortea.i@gmail.com>"]
readme = "README.md"
repository = "https://github.com/python-poetry/poetry"


[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.85.0"
FastAPI-SQLAlchemy = "^0.2.1"
pydantic = "^1.10.2"
alembic = "^1.8.1"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.3"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
graphene = "^3.1.1"
strawberry-graphql = {extras = ["fastapi"], version = "^0.133.2"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
black = "^22.8.0"
flake8 = "^5.0.4"
mypy = "^0.981"
pylint = "^2.15.3"
ipython = "^8.5.0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
